use Facade instead of Interface
