+++ src/main/java/org/betonquest/betonquest/api/feature/FeatureAPI.java
@@ -4,6 +4,7 @@ import org.betonquest.betonquest.config.QuestCanceler;
 import org.betonquest.betonquest.conversation.ConversationData;
 import org.betonquest.betonquest.id.ConversationID;
 import org.betonquest.betonquest.id.QuestCancelerID;
+import org.betonquest.betonquest.quest.registry.QuestRegistry;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
@@ -11,7 +12,21 @@ import java.util.Map;
 /**
  * The Feature logic.
  */
-public interface FeatureAPI {
+public final class FeatureAPI {
+
+    /**
+     * Quest Registry providing processors.
+     */
+    private final QuestRegistry questRegistry;
+
+    /**
+     * Create a new Feature API.
+     *
+     * @param questRegistry the registry containing processors
+     */
+    public FeatureAPI(final QuestRegistry questRegistry) {
+        this.questRegistry = questRegistry;
+    }
 
     /**
      * Gets stored Conversation Data.
@@ -19,15 +34,20 @@ public interface FeatureAPI {
      * The conversation data can be null if there was an error loading it.
      *
      * @param conversationID package name, dot and name of the conversation
-     * @return ConversationData object for this conversation or null if it does not exist
+     * @return ConversationData object for this conversation or null if it does
+     * not exist
      */
     @Nullable
-    ConversationData getConversation(ConversationID conversationID);
+    public ConversationData getConversation(final ConversationID conversationID) {
+        return questRegistry.conversations().getConversation(conversationID);
+    }
 
     /**
      * Get the loaded Quest Canceller.
      *
      * @return quest cancellers in a new map
      */
-    Map<QuestCancelerID, QuestCanceler> getCanceler();
+    public Map<QuestCancelerID, QuestCanceler> getCanceler() {
+        return questRegistry.questCanceller().getCancelers();
+    }
 }
